- Changed type in EncryptHeader from short to byte - Removed encrypt_entire_msg boolean and folded it into type - https://issues.jboss.org/browse/JGRP-1722

+++ tests/junit-functional/org/jgroups/tests/SizeTest.java
@@ -701,9 +701,9 @@ public class SizeTest {
 
 
     public static void testEncryptHeader() throws Exception {
-        ENCRYPT.EncryptHeader hdr=new ENCRYPT.EncryptHeader((short)1, null);
+        ENCRYPT.EncryptHeader hdr=new ENCRYPT.EncryptHeader((byte)1, null);
         _testSize(hdr);
-        hdr=new ENCRYPT.EncryptHeader((short)2, "Hello world");
+        hdr=new ENCRYPT.EncryptHeader((byte)2, "Hello world");
         _testSize(hdr);
     }
